refactor: move build sort key function to a more general place to get used by other components
