+++ data_types/Cargo.toml
@@ -12,6 +12,7 @@ croaring = "1.0.0"
 influxdb-line-protocol = { path = "../influxdb_line_protocol" }
 iox_time = { path = "../iox_time" }
 generated_types = { path = "../generated_types" }
+hashbrown.workspace = true
 observability_deps = { path = "../observability_deps" }
 once_cell = "1"
 ordered-float = "4"
